swimming/requirements.txt
anaconda-client==1.11.0
anaconda-navigator==2.3.1
appdirs==1.4.4
asgiref==3.6.0
atomicwrites==1.4.0
backports.weakref==1.0.post1
bkcharts==0.2
brotlipy==0.7.0
clyent==1.2.2
conda==23.1.0
conda-build==3.22.0
conda-repo-cli==1.0.20
conda-verify==3.4.2
constantly==15.1.0
cssselect==1.1.0
cytoolz==0.11.
daal4py==2021.6.0
datashape==0.5.4
Django==4.1.7
django-extensions==3.2.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.25.0
graphviz==0.20.1
inflection==0.5.1
llvmlite==0.38.0
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
munkres==1.1.4
mypy-extensions==0.4.3
navigator-updater==0.3.0
openpyxl==3.0.10
pandas==1.4.4
patsy==0.5.2
pep8==1.7.1
Pillow==9.2.0
ply==3.11
pyasn1-modules==0.2.8
pycosat==0.6.3
pycurl==7.45.1
PyDispatcher==2.0.5
pydot==1.4.2
pyls-spyder==0.4.0
PyQt5-sip==12.11.0
pytest==7.1.2
python-lsp-jsonrpc==1.0.0
PyYAML==6.0
queuelib==1.5.0
scikit-learn-intelex==2021.20221004.121333
scipy==1.9.1
sqlparse==0.4.3
TBB==0.2
torch==1.12.1
webencodings==0.5.1
zict==2.1.0
